pyproject.toml
[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[project]
name = "hostfactory"
description = "Symphony k8s deployment tools"

readme = "README.md"
requires-python = ">=3.12"

# dynamic = ["version"]
version = "1.0.0"

dependencies = [
    "click>=8",
    "kubernetes",
    "boto3",
    "jinja2",
    "typing-extensions",
    "inotify",
    "hatchling",
    "rich",
    "wrapt",
    "pydantic",
    "pydantic-settings",
    "tenacity",
    "prometheus-client",
    "sqlalchemy",
    "psycopg2",
    "alembic",
    "fastapi",
    "uvicorn"
]

[project.optional-dependencies]
testing = [
    "PyYAML",
    "coverage",
    "pytest-cov",
    "pytest",
    "mypy",
    "ruff==0.12.2",
    "typos",
    "httpx"
]
documentation = [
    "sphinx",
    "sphinx-click",
    "myst-parser",
]

[project.scripts]
hostfactory = "hostfactory.cli.hf:run"
hostfactory-admin = "hostfactory.cli.hfadmin:run"
hostfactory-events = "hostfactory.cli.events_db:run"
hostfactory-utils = "hostfactory.cli.hfutils:runserver"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.sdist]
exclude = ["/*venv"]

[tool.pytest.ini_options]
addopts = "--cov src/hostfactory --cov-report term-missing --verbose --cov-fail-under 40"
norecursedirs = ["dist", "build", ".tox", "relvenv"]
testpaths = ["src/hostfactory/tests/unit","src/hostfactory/test/regression"]
markers = [
    "unit: testing with no external dependencies",
    "regression: expects the pr instance to be up"]

[tool.coverage.run]
branch = true
omit = ["src/hostfactory/tests/*"]

[tool.coverage.paths]
source = [
    "src/hostfactory/",
    "*/site-packages/",
]

[tool.coverage.report]
# Regexes for lines to exclude from consideration
exclude_also = [
    # Don't complain about missing debug-only code:
    "def __repr__",
    "if self\\.debug",
    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",
    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",
    # Don't complain about abstract methods, they aren't run:
    "@(abc\\.)?abstractmethod",
]

[tool.mypy]
# If we ever refactor, we need to deal with this "no_implicit_optional"
no_implicit_optional = false
no_namespace_packages = true
warn_return_any = true
# Remove "import-untyped" when installing new packages in case it has a type hints package, remove the rest as we refactor.
disable_error_code = ["call-arg", "attr-defined", "arg-type", "valid-type", "import-untyped", "union-attr", "assignment", "no-any-return", "return-value", "import-not-found"]

[tool.uv]

[tool.uv.sources]

[tool.typos]
# The signer API contains certificate data and appears to confuse the spellchecker
files.extend-exclude = ["*.json"]

[tool.typos.default.extend-words]
# $nin is a mongodb query [not in], 
# requestor is valid as it is in reference to a technical interface [https://english.stackexchange.com/questions/29254/whats-the-difference-between-requester-and-requestor]
requestor = "requestor"
nin = "nin"

